% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basestats.R
\name{basestats}
\alias{basestats}
\title{Creates a basic summary of the raw data and density plot raw data}
\usage{
basestats(df)
}
\arguments{
\item{df}{The data frame to summarise}
}
\value{
plot with density for raw data
}
\description{
Creates a basic summary of the raw data and density plot raw data
}
\examples{
basestats(celegance)

}
