% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosisresponse.R
\name{dosisresponse}
\alias{dosisresponse}
\title{This function takes a list of chemical components and returns a plot with dose response comparison for each selected component}
\usage{
dosisresponse(df, components)
}
\arguments{
\item{df}{The data frame to summarise}

\item{components}{the charakter vector of components to be used in analyse}
}
\value{
plot
}
\description{
This function takes a list of chemical components and returns a plot with dose response comparison for each selected component
}
\examples{

components <-c("2,6-diisopropylnaphthalene","S-medium","naphthalene", "decane")
dosisresponse(celgenace,components)


}
